+++ examples/fulldemo/plugins/repeat_demo/__init__.py
@@ -13,7 +13,7 @@ from flask import abort, jsonify
 
 __plugin_name__ = "repeatdemo"
 __author__ = "Mr.tao <staugur@saintic.com>"
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 
 def view_abort_403():
@@ -21,13 +21,42 @@ def view_abort_403():
 
 
 def permission_deny(error):
-    return jsonify(dict(status=403, msg="permission deny")),403
+    return jsonify(dict(status=403, msg="permission deny")), 403
+
+
+class ApiError(Exception):
+
+    def __init__(self, code, message, status_code=200):
+        super(ApiError, self).__init__()
+        self.code = code
+        self.message = message
+        self.status_code = status_code
+
+    def to_dict(self):
+        rv = dict(code=self.code, msg=self.message)
+        return rv
+
+
+def handle_api_error(e):
+    response = jsonify(e.to_dict())
+    response.status_code = e.status_code
+    return response
+
+
+def raise_api_error_view():
+    raise ApiError(10000, "test_err_class_handler")
 
 
 def register():
     return {
-        'vep': dict(rule='/403', view_func=view_abort_403),
+        'vep': [
+            dict(rule='/403', view_func=view_abort_403),
+            dict(rule='/api_error', view_func=raise_api_error_view)
+        ],
         'filter': dict(repeat_filter=lambda x: 'test-filter-repeat'),
-        'errhandler': {403: permission_deny},
+        'errhandler': [
+            dict(error=403, handler=permission_deny),
+            dict(error=ApiError, handler=handle_api_error)
+        ],
         'tcp': dict(change_to_str=str),
     }